Simplify API entrypoint
